refactor a big file into 3 files
